Introduce game board data structure
To clean up the gameplay utility function definitions, the char array
that stores the game board and the mines are now part of a single
struct, which stores the dimensions of the board as well as a pointer to
the array that represents the board. The memory for the board is
dynamically allocated in order to allow for more flexibility in custom
game board dimension in the future.

The code for working with board arrays has been modified to work with
the Board struct.

board.h and board.cpp
These files contain the declaration of the Board struct, as well as
declarations and definitions for the functions to allocate and free
memory for their arrays.