Remove C++ default function parameters
The game no longer relies on default parameters in functions. Instead,
multiple functions with similar names but different parameters have been
defined, in preparation for the code to be converted to C.

Comments describing the functions defined in gamefunctions.cpp were
moved or added to their declarations in the corresponding header file.
Additionally, the labels for each parameter are also defined in the
header file, to aid in knowing which argument goes where.