Add savegame functionality
This commit introduces a lot of changes in the entire project, so let's
go file by file. Most of these changes revolve around the creation of
the Savegame data structure.

game.c
The game function was moved to its own file to improve readability,
since it is functionally different enough from main to deserve its own
file. A fourth argument was added, which is a pointer to a Savegame data
structure. The game compares this pointer against NULL; if the pointer
is NULL, then the Savefile object is not read from. If the pointer is
not null, then the structure is read from to load the game.

savegame.c and savegame.h
Define the Savegame data structure, as well as 4 utility functions for
I/O using the structure.

board.c
memset is now used to initialize alll Board structures. This way, we
don't have to use for loops in the game function every time we want to
initialize a Board structure. memset should also be marginally faster
than using a for loop to iterate through the whole array.

gamefunctions.c
The internal definition for a mine board has changed in order to make
the code simpler. Whereas previously a '.' would signify an empty
square, now a '+' represents an empty square. This was done so that the
memset statement in initBoardArray could apply to both types of boards.
A new pause menu entry was also added for saving the game.

main.c
Added a screen that prompts the user to either select a New game or to
load the saved game from disk.