Write generic menu function
There were so many instances of multiple choice user prompts in the code
that it just made more sense to write a generic menu function to prompt
the user to select between a variable number of options. The function is
variadic, so it can be called with any number options. Technically, even
calling with 0 options is possible, but it isn't recommened. Calling
with 1 option is the equivalent of an "OK" prompt.