Remove need for two Board structs
Technically, I never actually needed 2 separate arrays to store board
data and mine data. But I did that anyways, and this is where I redeemed
myself. The MSB of every character in the board struct's array is set if
there is a mine at that coordinate. The other 7 bits are used to store
the character representation of that coordinate.