Delete savefile if player dies
The savefile is only deleted if the player died IN THAT GAME. This is as
opposed to dying in a new game even though you technically have a
savefile on your hard drive. If you load a saved game and die in that
game, you will not be able to try again at that particular game.